The parity test caught a real one: float32 in the RNG
First run in Unreal: every hex section clean, hashString clean, seedPhrase
clean, and — the part that matters — all 250 raw 32-bit draws from all six
streams clean. The generator is bit-exact. 117 failures, all in one sub-check.

    worldgen float draw 0: got 6.4754543304443359, expected 6.4754541730508208

That gap is 1.6e-7, which is float32. FloatRange computed in double and then
narrowed on the way out:

    float FloatRange(float Min, float Max)
    { return static_cast<float>(Min + NextDouble() * (Max - Min)); }

Math.fround(6.475454173050821) is 6.475454330444336 — the reported number to
the digit. So the fix is the type, not the tolerance: the whole floating-point
surface of the RNG is now double, matching JS, which has no other kind of
number. Chance took a double too — passing 0.3f meant testing against
0.30000001192092896, which held for these 40 draws by luck rather than design.

Same hazard removed from FHexPath::Cost and FHexReach::Cost. Those passed only
because every cost in the vectors is a small integer and survives float32
intact; a fractional terrain cost would have started drifting later, far from
anything that would point here.

Recorded why verify-port.mjs slept through it: JavaScript has one number type,
so the mirror modelled `float` as double and agreed with a C++ that was wrong.
A narrowing can only be represented there with Math.fround at the same point.
Noted in the header, since the harness is otherwise good enough to trust.

+++ ue-port/Source/LandnamHex.cpp
@@ -259,7 +259,7 @@ FHexPath UHexLib::FindPathNative(const FHex& Start, const FHex& Goal, TFunctionR
 	if (Start == Goal)
 	{
 		Result.Hexes.Add(Start);
-		Result.Cost = 0.0f;
+		Result.Cost = 0.0;
 		Result.bReachable = true;
 		return Result;
 	}
@@ -293,7 +293,7 @@ FHexPath UHexLib::FindPathNative(const FHex& Start, const FHex& Goal, TFunctionR
 			Algo::Reverse(Reversed);
 
 			Result.Hexes = MoveTemp(Reversed);
-			Result.Cost = static_cast<float>(GScore[Goal]);
+			Result.Cost = GScore[Goal];
 			Result.bReachable = true;
 			return Result;
 		}
@@ -374,10 +374,10 @@ FHexPath UHexLib::FindPath(const FHex& Start, const FHex& Goal, const FHexCostDe
 	return FindPathNative(Start, Goal, [&Cost](const FHex& H) { return CostFromDelegate(Cost, H); });
 }
 
-TArray<FHexReach> UHexLib::Reachable(const FHex& Start, float Budget, const FHexCostDelegate& Cost)
+TArray<FHexReach> UHexLib::Reachable(const FHex& Start, double Budget, const FHexCostDelegate& Cost)
 {
 	const TMap<FHex, double> Best = ReachableNative(
-		Start, static_cast<double>(Budget),
+		Start, Budget,
 		[&Cost](const FHex& H) { return CostFromDelegate(Cost, H); });
 
 	TArray<FHexReach> Out;
@@ -386,7 +386,7 @@ TArray<FHexReach> UHexLib::Reachable(const FHex& Start, float Budget, const FHex
 	{
 		FHexReach Entry;
 		Entry.Hex = Pair.Key;
-		Entry.Cost = static_cast<float>(Pair.Value);
+		Entry.Cost = Pair.Value;
 		Out.Add(Entry);
 	}
 	return Out;
